rename jfr module
