+++ instrumentation/runtime-telemetry-jfr/javaagent/build.gradle.kts
@@ -1,0 +1,12 @@
+plugins {
+  id("otel.javaagent-instrumentation")
+}
+otelJava {
+  minJavaVersionSupported.set(JavaVersion.VERSION_17)
+}
+
+dependencies {
+  implementation(project(":instrumentation:runtime-telemetry-jfr:library"))
+
+  compileOnly("io.opentelemetry:opentelemetry-sdk-extension-autoconfigure")
+}
