+++ instrumentation/runtime-telemetry-jfr/javaagent/src/test/groovy/io/opentelemetry/instrumentation/javaagent/runtimetelemetryjfr/RuntimeMetricsInstallerJfr.groovy
