+++ settings.gradle.kts
@@ -423,8 +423,8 @@ hideFromDependabot(":instrumentation:rocketmq:rocketmq-client:rocketmq-client-5.
 hideFromDependabot(":instrumentation:rocketmq:rocketmq-client:rocketmq-client-5.0:testing")
 hideFromDependabot(":instrumentation:runtime-metrics:javaagent")
 hideFromDependabot(":instrumentation:runtime-metrics:library")
-hideFromDependabot(":instrumentation:runtime-metrics-jfr:javaagent")
-hideFromDependabot(":instrumentation:runtime-metrics-jfr:library")
+hideFromDependabot(":instrumentation:runtime-telemetry-jfr:javaagent")
+hideFromDependabot(":instrumentation:runtime-telemetry-jfr:library")
 hideFromDependabot(":instrumentation:rxjava:rxjava-1.0:library")
 hideFromDependabot(":instrumentation:rxjava:rxjava-2.0:library")
 hideFromDependabot(":instrumentation:rxjava:rxjava-2.0:testing")
